Added return type MetaNearestStationResult to also give distance in km
