+++ src/simultaneousness_analysis/meta/search.py
@@ -31,3 +31,28 @@ class MetaSearch:
         init=False,
         default_factory=lambda: FILTER_FUNCTIONS,
     )
+
+
+@dataclass(frozen=True, kw_only=True)
+class MetaSearchResult:
+    """Object containing the results of a metadata search and the search parameters used for the search."""
+
+    search: MetaSearch
+    paths: list[str] | None = None
+
+    @property
+    def length(self) -> int:
+        """Returns the number of paths in the search results.
+
+        Returns:
+            int: Number of paths in the search results.
+        """
+        return len(self.paths) if self.paths is not None else 0
+
+
+@dataclass(frozen=True, kw_only=True)
+class MetaNearestStationResult:
+    """Nearest station lookup result containing station id and distance."""
+
+    stations_id: int
+    distance_km: float
